Fix signed vs unsigned comparison warning on Windows

+++ libtiff/tif_dirread.c
@@ -862,7 +862,7 @@ static enum TIFFReadDirEntryErr TIFFReadDirEntryArrayWithLimit(
 	datasize=(*count)*typesize;
 	assert((tmsize_t)datasize>0);
 
-	if( isMapped(tif) && datasize > tif->tif_size )
+	if( isMapped(tif) && datasize > (uint32)tif->tif_size )
 		return TIFFReadDirEntryErrIo;
 
 	if( !isMapped(tif) &&
